backend/requirements.txt
# ── Flask 存量服务依赖 ─────────────────────────────────────────────────────────
flask==3.1.0
flask-socketio==5.4.1
python-socketio==5.11.3
eventlet==0.38.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-cors==5.0.1
flask-bcrypt==1.0.1
flask-migrate==4.0.7
flask-limiter==3.8.0
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==44.0.3
gunicorn==22.0.0
redis==5.2.1
openpyxl==3.1.5
apscheduler>=3.10.4,<4.0

# ── FastAPI 新增模块依赖 ───────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
PyJWT==2.10.1
httpx==0.28.1

# ── 测试 ──────────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-flask==1.3.0
